- switched to custom label drawing

+++ NimbleCommander/States/FilePanels/PanelBriefViewCollectionViewLayout.mm
@@ -23,14 +23,14 @@
 - (NSArray<__kindof NSCollectionViewLayoutAttributes *> *)layoutAttributesForElementsInRect:(NSRect)rect
 {
     NSArray<__kindof NSCollectionViewLayoutAttributes *> *attrs = [super layoutAttributesForElementsInRect:rect];
+    
     const int items_per_column = self.rowsCount;
     
     struct Col {
-        short width = 0;
-        short origin = numeric_limits<short>::max();
+        int width = 0;
+        int origin = numeric_limits<int>::max();
     };
     vector<Col> columns;
-    
     
     for( NSCollectionViewLayoutAttributes *i in attrs ) {
         const int index = (int)i.indexPath.item;
@@ -40,6 +40,7 @@
         
         if( col >= columns.size() )
             columns.resize(col+1);
+        
         
         if( columns[col].width < orig_frame.size.width )
             columns[col].width = orig_frame.size.width;
@@ -54,24 +55,16 @@
         const int col = index / items_per_column;
         NSRect orig_frame = i.frame;
         
-        //        int x_origin = 0;
-        //        for( auto n = 0; n < col; ++n ) x_origin += column_widths[n];
-        
-        
-        //NSRect new_frame = NSMakeRect(orig_frame.origin.x,
         NSRect new_frame = NSMakeRect(columns[col].origin,
-                                      //                                      row * (self.itemSize.height + self.minimumInteritemSpacing),
                                       row * item_height,
                                       columns[col].width,
                                       item_height);
         
         i.frame = new_frame;
-        
     }
-    
+
     return attrs;
-    
-    
 }
+
 @end
